+++ src/components/common/EmptyContent/index.tsx
@@ -1,0 +1,3 @@
+import EmptyContent from '@/components/common/EmptyContent/EmptyContent';
+
+export { EmptyContent };
