+++ src/components/common/index.ts
@@ -1,5 +1,6 @@
 import { Button } from '@/components/common/Button';
 import { Calendar } from '@/components/common/Calendar';
+import { EmptyContent } from '@/components/common/EmptyContent';
 import { RootErrorFallback } from '@/components/common/ErrorBoundary';
 import { Form } from '@/components/common/Form';
 import { Loading } from '@/components/common/Loading';
@@ -7,4 +8,4 @@ import { Overlay } from '@/components/common/Overlay';
 
 export * from './Header';
 export * from './Toast';
-export { Button, Calendar, Form, Loading, Overlay, RootErrorFallback };
+export { Button, Calendar, EmptyContent, Form, Loading, Overlay, RootErrorFallback };
